コンポーネントのリファクタリングとVideoDisplayの追加
