+++ src/components/webcam.tsx
@@ -1,25 +1,13 @@
 import React, { useEffect, useRef, useState, useCallback } from 'react'
 
 import homeStore from '@/features/stores/home'
-import settingsStore from '@/features/stores/settings'
-import { IconButton } from './iconButton'
-import { useDraggable } from '@/hooks/useDraggable'
+import { VideoDisplay } from './common/VideoDisplay'
 
 export const Webcam = () => {
-  const triggerShutter = homeStore((s) => s.triggerShutter)
-  const useVideoAsBackground = settingsStore((s) => s.useVideoAsBackground)
   const [selectedDevice, setSelectedDevice] = useState<string>('')
   const [devices, setDevices] = useState<MediaDeviceInfo[]>([])
   const [showRotateButton, setShowRotateButton] = useState(true)
-  const [isExpanded, setIsExpanded] = useState(false)
   const videoRef = useRef<HTMLVideoElement>(null)
-  const backgroundVideoRef = useRef<HTMLVideoElement>(null)
-  const {
-    isMobile,
-    handleMouseDown,
-    resetPosition,
-    style: dragStyle,
-  } = useDraggable()
 
   const refreshDevices = useCallback(async () => {
     if (!navigator.mediaDevices) return
@@ -51,18 +39,6 @@ export const Webcam = () => {
     }
   }, [refreshDevices])
 
-  useEffect(() => {
-    if (useVideoAsBackground && videoRef.current?.srcObject) {
-      if (backgroundVideoRef.current) {
-        backgroundVideoRef.current.srcObject = videoRef.current.srcObject
-      }
-    } else if (!useVideoAsBackground) {
-      if (backgroundVideoRef.current) {
-        backgroundVideoRef.current.srcObject = null
-      }
-    }
-  }, [useVideoAsBackground])
-
   const initializeCamera = useCallback(async () => {
     if (!navigator.mediaDevices || !selectedDevice) return
     try {
@@ -73,13 +49,10 @@ export const Webcam = () => {
       if (videoRef.current) {
         videoRef.current.srcObject = stream
       }
-      if (backgroundVideoRef.current && useVideoAsBackground) {
-        backgroundVideoRef.current.srcObject = stream
-      }
     } catch (e) {
       console.error('Error initializing camera:', e)
     }
-  }, [selectedDevice, useVideoAsBackground])
+  }, [selectedDevice])
 
   useEffect(() => {
     initializeCamera()
@@ -100,91 +73,13 @@ export const Webcam = () => {
     initializeCamera()
   }, [selectedDevice, initializeCamera])
 
-  const handleCapture = useCallback(() => {
-    const canvas = document.createElement('canvas')
-    canvas.width = videoRef.current!.videoWidth
-    canvas.height = videoRef.current!.videoHeight
-    const ctx = canvas.getContext('2d')
-    if (!ctx) return
-    ctx.drawImage(videoRef.current!, 0, 0)
-    const data = canvas.toDataURL('image/png')
-
-    if (data !== '') {
-      console.log('capture')
-      homeStore.setState({
-        modalImage: data,
-        triggerShutter: false, // シャッターをリセット
-      })
-    } else {
-      homeStore.setState({ modalImage: '' })
-    }
-  }, [])
-
-  useEffect(() => {
-    if (triggerShutter) {
-      handleCapture()
-    }
-  }, [triggerShutter, handleCapture])
-
-  const handleExpand = useCallback(() => {
-    setIsExpanded(!isExpanded)
-    // When expanded, use video as background
-    settingsStore.setState({ useVideoAsBackground: !isExpanded })
-    // Reset position and size when toggling background mode
-    resetPosition()
-  }, [isExpanded, resetPosition])
-
   return (
-    <>
-      {useVideoAsBackground && (
-        <video
-          ref={backgroundVideoRef}
-          autoPlay
-          playsInline
-          muted
-          className="fixed top-0 left-0 w-full h-full object-cover -z-10"
-        />
-      )}
-      <div className="fixed right-4 top-4 max-h-[40vh] z-10" style={dragStyle}>
-        <div
-          className="relative w-full md:max-w-[512px] max-w-[70%] select-none"
-          onMouseDown={!isMobile ? handleMouseDown : undefined}
-        >
-          <video
-            ref={videoRef}
-            width={512}
-            height={512}
-            id="local-video"
-            autoPlay
-            playsInline
-            muted
-            className={`rounded-lg w-auto object-contain max-h-[100%] ${
-              useVideoAsBackground ? 'invisible' : ''
-            }`}
-          />
-          <div className="md:block absolute top-2 right-2">
-            <IconButton
-              iconName="24/Roll"
-              className="bg-secondary hover:bg-secondary-hover active:bg-secondary-press disabled:bg-secondary-disabled m-2"
-              isProcessing={false}
-              disabled={!showRotateButton}
-              onClick={handleRotateCamera}
-            />
-            <IconButton
-              iconName="24/Expand"
-              className="bg-secondary hover:bg-secondary-hover active:bg-secondary-press disabled:bg-secondary-disabled m-2"
-              isProcessing={false}
-              onClick={handleExpand}
-            />
-            <IconButton
-              iconName="24/Shutter"
-              className="z-30 bg-secondary hover:bg-secondary-hover active:bg-secondary-press disabled:bg-secondary-disabled m-2"
-              isProcessing={false}
-              onClick={handleCapture}
-            />
-          </div>
-        </div>
-      </div>
-    </>
+    <VideoDisplay
+      videoRef={videoRef}
+      onToggleSource={handleRotateCamera}
+      toggleSourceIcon="24/Roll"
+      toggleSourceDisabled={!showRotateButton}
+      showToggleButton={true}
+    />
   )
 }
